Require actual text in the JIRA issue key
The * in the regex was matching an empty string, which would allow
something like [- fooo] meaning a jira issue, and resulting in trying to
look up a jira issue with the key '-', which I'm pretty sure is not what
we want.  Require some content on both sides of the dash.

Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>