Stop processing comments when meeting a non-valid JIRA issue key tag
This allows having the progress in top of the status file and other
content further down in the file.

For example, the following now works:

[VIRT-123]
Some comment to upload to JIRA.

[VIRT-124 # This is a fancy virt issue]
Some comment to upload to JIRA.

[Plans]
 * Something that doesn't go to JIRA

[VIRT-125]
Some comment to upload to JIRA.

[Other]
Nothing else for JIRA.

This should solve issue #11 as well.

Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>