Add -s switch to not show entire status
Some people, like myself, keep a lot of random notes in my weekly
todo/status file, that I don't need repeated on the console, because
I'll have to scroll back up to see the jira updates.

Provide a switch -s, for silent, so that only the JIRA updates are shown
and not the entire status file.

Signed-off-by: Christoffer Dall <christoffer.dall@linaro.org>