Adding new FIN token
Adding a new FIN token to tell the parser to stop processing entirely
and that anything below the token should be ignored.

That way it is possible to include a "Plans" section that has the same
format as in the highlight but describe what one is planning to do for
the coming week.  Up to now the only way to do this was to  add a '#'
character at the beginning of each line.

Signed-off-by: Mathieu Poirier <mathieu.poirier@linaro.org>
Signed-off-by: Christoffer Dall <cdall@cs.columbia.edu>