jiralogin: open temporary file in text mode
tempfile.NamedTemporaryFile will open a temporary file in byte mode by
default. However what we need is a text file. When running with
Python3 it throws this error:

Traceback (most recent call last):
  File "./jipdate.py", line 458, in <module>
    main(sys.argv)
  File "./jipdate.py", line 442, in main
    filename = get_jira_issues(jira, username)
  File "./jipdate.py", line 211, in get_jira_issues
    f.write(subject)
  File "/usr/lib/python3.7/tempfile.py", line 481, in func_wrapper
    return func(*args, **kwargs)
TypeError: a bytes-like object is required, not 'str'

So we need to ensure the file is opened in text mode.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>