jipdate: add support for setting 'resolution' when resolving
When doing a transition to Resolved, jira will by default set the
resolution to 'Delivered'. This patch adds support to change
that. To set resolution, during a transition to 'Resolved', we need to
use the following 'Status' pattern:

Status: Resolved / <resolution>

Where <resolution> can be Cancelled, Deffered, Out of scope, .. or any
other valid resolution.

Like with transition, we are querying Jira for the list of valid
resolutions (jira.resolution()) and we reject invalid
resolution. During the parsing all extra spaces are removed, and words
are capitalized.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>