jipdate: capitalize transition names
When we do a state transition, we first query Jira for all possible
states, and then later we do a string comparison. So far we have
assumed that Jira is using 'capitalized' words and we use
str.title(). However Jira doesn't really capitalize words like
str.title(), especially when there is a ' in the word. e.g.

>>> str.title("I'm a state")
"I'M A State"

This was found looking at resolution field such as "Won't do".

So to make sure that strings are capitalized properly before we do the
comparison, let's add a str.title() on the strings returned by Jira.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>