jipdate: avoid refetching issues during the parsing of the text file
During the parsing of the text file, we re-fetch each issue one by
one, and it can take a lot of time, when processing many issues at
once. We don't need to do that since we have already fetched all
issues when we generated the file. So this patch keeps track of
'my_issues' array and pass it around. In the parsing code, we just
need to locate the right issue object.

The parsing is significantly faster with this change.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>