jira: add url attribute to server details
This is a preparation patch to add support for different
authentication, such as token based authentication. We want to be able
to define if a server supports password or token authentication, so in
a next patch, we will add more attributes.

For now, we only use 'url'. This patch is no-op.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>