cfg: add 'server' in user config file
Allow the user to specify a server (url) in the config file. If
defined, then it will be used, otherwise we use the default value from
cfg.py.

Since the user can now use any server, we can no longer display the
"OFFICIAL" vs "TEST" message we've had so far. Instead we show the URL
we are about to publish to. It should be sufficient to warn users.

Signed-off-by: Nicolas Dechesne <nicolas.dechesne@linaro.org>