jiralogin: report correct server to update
The server to connect to was correctly read out for the connection
itself, but later on when asking for confirmation it reported the wrong
server. This because the server in the cfg file was set to the (old)
production server and didn't query the setting in the yaml file.

To fix this we've wrapped the retrieval of the server details in a
single function, which now selects the correct server.

The test server has been kept for now, but should be removed since we
won't have a test server in the future.

Fixes #44.

Signed-off-by: Joakim Bech <joakim.bech@linaro.org>