jipsearch: Add "-s FIELD:VALUE" parameter
... to fetch and print a single field for matched cards.

This option is useful for fetching specific information about
cards that is not part of default printout.  E.g., to fetch
parent of a specific card one can use:
$ jipsearch -j "key=GNU-835" -s parent:key
And to fetch assignee of that card:
$ jipsearch -j "key=GNU-835" -s assignee:emailAddress

This is used in Toolchain WG CI to automatically create sub-task
cards for detected regressions and assign them to this week's
CI babysitter.

Signed-off-by: Maxim Kuvyrkov <maxim.kuvyrkov@linaro.org>
Signed-off-by: Anders Roxell <anders.roxell@linaro.org>